+++ src/wirestate/alias.ts
@@ -1,5 +1,4 @@
 import { LazyServiceIdentifier, type ServiceIdentifier } from "inversify";
-import { action, computed, observable } from "mobx";
 
 export {
   Container,
@@ -15,49 +14,8 @@ export {
   type ServiceIdentifier,
   bindingTypeValues as BindingType,
   bindingScopeValues as ScopeBindingType,
+  LazyServiceIdentifier,
 } from "inversify";
-export {
-  autorun,
-  comparer,
-  configure,
-  flow,
-  flowResult,
-  isAction,
-  isFlow,
-  isFlowCancellationError,
-  isObservable,
-  makeAutoObservable,
-  makeObservable,
-  reaction,
-  runInAction,
-  toJS,
-  when,
-} from "mobx";
-export { observer } from "mobx-react-lite";
-
-export function Observable() {
-  return observable;
-}
-
-export function ShallowObservable() {
-  return observable.shallow;
-}
-
-export function RefObservable() {
-  return observable.ref;
-}
-
-export function DeepObservable() {
-  return observable.deep;
-}
-
-export function Action() {
-  return action;
-}
-
-export function Computed() {
-  return computed;
-}
 
 export function forwardRef<TInstance = unknown>(forward: () => ServiceIdentifier<TInstance>) {
   return new LazyServiceIdentifier(forward);