+++ src/wirestate/core/container/query-optional.ts
@@ -5,7 +5,7 @@ import { prefix } from "@/macroses/prefix.macro";
 
 import { QueryBus } from "@/wirestate/core/queries/query-bus";
 import { QUERY_BUS_TOKEN } from "@/wirestate/core/registry";
-import type { MaybePromise } from "@/wirestate/types/general";
+import type { MaybePromise, Optional } from "@/wirestate/types/general";
 import type { TQueryType } from "@/wirestate/types/queries";
 
 /**
@@ -20,7 +20,7 @@ export function queryOptional<R = unknown, D = unknown>(
   container: Container,
   type: TQueryType,
   data?: D
-): MaybePromise<R> | null {
+): Optional<MaybePromise<R>> {
   dbg.info(prefix(__filename), "Optional query data:", type, data, container);
 
   return container.get<QueryBus>(QUERY_BUS_TOKEN).queryOptional<R, D>(type, data);