+++ src/wirestate/core/queries/use-optional-query-caller.ts
@@ -8,7 +8,7 @@ import { useContainer } from "@/wirestate/core/provision/use-container";
 import { QueryBus } from "@/wirestate/core/queries/query-bus";
 import { QUERY_BUS_TOKEN } from "@/wirestate/core/registry";
 import type { MaybePromise, Optional } from "@/wirestate/types/general";
-import type { TQueryType } from "@/wirestate/types/queries";
+import type { TOptionalQueryCaller, TQueryType } from "@/wirestate/types/queries";
 
 /**
  * Returns a function to dispatch optional queries on the active container.
@@ -16,17 +16,21 @@ import type { TQueryType } from "@/wirestate/types/queries";
  *
  * @returns optional query dispatcher
  */
-export function useOptionalQueryCaller() {
+export function useOptionalQueryCaller<
+  R = unknown,
+  D = unknown,
+  T extends TQueryType = TQueryType,
+>(): TOptionalQueryCaller<R, D, T> {
   const container: Container = useContainer();
 
   return useCallback(
-    <R = unknown, D = unknown>(type: TQueryType, data?: D): Optional<MaybePromise<R>> => {
+    (type: T, data?: D): Optional<MaybePromise<R>> => {
       dbg.info(prefix(__filename), "Optional query data:", {
         type,
         data,
       });
 
-      return container.get<QueryBus>(QUERY_BUS_TOKEN).queryOptional<R, D>(type, data);
+      return container.get<QueryBus>(QUERY_BUS_TOKEN).queryOptional<R, D, T>(type, data);
     },
     [container]
   );